move inventory2 to forseti
