+++ google/cloud/security/common/gcp_type/resources.py
