+++ google/cloud/security/common/storage/base_storage.py
