+++ google/cloud/security/common/storage/sql_storage.py
@@ -22,12 +22,10 @@ from sqlalchemy import Text
 from sqlalchemy import DateTime
 from sqlalchemy import Integer
 from sqlalchemy import and_
+from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import aliased
 
-
-from sqlalchemy.ext.declarative import declarative_base
-from google.cloud.security.iam.inventory.inventory2.storage import \
-    Storage as BaseStorage
+from google.cloud.security.common.storage.base_storage import Storage as BaseStorage
 
 # TODO: Remove this when time allows
 # pylint: disable=missing-type-doc,missing-return-type-doc,missing-return-doc
