+++ tests/iam/inventory_tests/storage_test.py
@@ -20,7 +20,7 @@ from sqlalchemy import event
 
 from google.cloud.security.iam.inventory.storage import Storage, initialize
 from google.cloud.security.iam import db
-from google.cloud.security.iam.inventory.inventory2.resources import Resource
+from google.cloud.security.common.gcp_type.resources import Resource
 from tests.iam.utils.db import create_test_engine
